IGNITE-28539 Rename PartitionReplicaListener and related classes
