+++ modules/table/src/test/java/org/apache/ignite/internal/table/distributed/replication/DefaultTablePartitionReplicaProcessorIndexLockingTest.java
@@ -31,9 +31,8 @@ import static org.apache.ignite.internal.partition.replicator.network.replicatio
 import static org.apache.ignite.internal.partition.replicator.network.replication.RequestType.RW_REPLACE_IF_EXIST;
 import static org.apache.ignite.internal.partition.replicator.network.replication.RequestType.RW_UPSERT;
 import static org.apache.ignite.internal.partition.replicator.network.replication.RequestType.RW_UPSERT_ALL;
-import static org.apache.ignite.internal.table.distributed.replication.PartitionReplicaListenerIndexLockingTest.LOCAL_NODE_ID;
-import static org.apache.ignite.internal.table.distributed.replication.PartitionReplicaListenerTest.binaryRowsToBuffers;
-import static org.apache.ignite.internal.table.distributed.replication.PartitionReplicaListenerTest.zonePartitionIdMessage;
+import static org.apache.ignite.internal.table.distributed.replication.DefaultTablePartitionReplicaProcessorTest.binaryRowsToBuffers;
+import static org.apache.ignite.internal.table.distributed.replication.DefaultTablePartitionReplicaProcessorTest.zonePartitionIdMessage;
 import static org.apache.ignite.internal.testframework.IgniteTestUtils.await;
 import static org.apache.ignite.internal.tx.TxState.checkTransitionCorrectness;
 import static org.apache.ignite.internal.util.CompletableFutures.nullCompletedFuture;
@@ -47,6 +46,7 @@ import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMaps;
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -85,6 +85,7 @@ import org.apache.ignite.internal.replicator.message.ReplicaRequest;
 import org.apache.ignite.internal.schema.AlwaysSyncedSchemaSyncService;
 import org.apache.ignite.internal.schema.BinaryRow;
 import org.apache.ignite.internal.schema.BinaryRowConverter;
+import org.apache.ignite.internal.schema.BinaryTupleSchema;
 import org.apache.ignite.internal.schema.Column;
 import org.apache.ignite.internal.schema.ColumnsExtractor;
 import org.apache.ignite.internal.schema.SchemaDescriptor;
@@ -93,17 +94,21 @@ import org.apache.ignite.internal.schema.marshaller.reflection.ReflectionMarshal
 import org.apache.ignite.internal.storage.RowId;
 import org.apache.ignite.internal.storage.impl.TestMvPartitionStorage;
 import org.apache.ignite.internal.storage.index.SortedIndexStorage;
+import org.apache.ignite.internal.storage.index.StorageHashIndexDescriptor;
+import org.apache.ignite.internal.storage.index.StorageHashIndexDescriptor.StorageHashIndexColumnDescriptor;
 import org.apache.ignite.internal.storage.index.StorageSortedIndexDescriptor;
 import org.apache.ignite.internal.storage.index.StorageSortedIndexDescriptor.StorageSortedIndexColumnDescriptor;
+import org.apache.ignite.internal.storage.index.impl.TestHashIndexStorage;
 import org.apache.ignite.internal.storage.index.impl.TestSortedIndexStorage;
 import org.apache.ignite.internal.table.TableTestUtils;
+import org.apache.ignite.internal.table.distributed.HashIndexLocker;
 import org.apache.ignite.internal.table.distributed.IndexLocker;
 import org.apache.ignite.internal.table.distributed.SortedIndexLocker;
 import org.apache.ignite.internal.table.distributed.StorageUpdateHandler;
 import org.apache.ignite.internal.table.distributed.TableSchemaAwareIndexStorage;
 import org.apache.ignite.internal.table.distributed.index.IndexMetaStorage;
 import org.apache.ignite.internal.table.distributed.index.IndexUpdateHandler;
-import org.apache.ignite.internal.table.distributed.replicator.PartitionReplicaListener;
+import org.apache.ignite.internal.table.distributed.replicator.DefaultTablePartitionReplicaProcessor;
 import org.apache.ignite.internal.table.impl.DummyInternalTableImpl;
 import org.apache.ignite.internal.table.impl.DummySchemaManagerImpl;
 import org.apache.ignite.internal.table.impl.DummyValidationSchemasSource;
@@ -135,13 +140,19 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 /** There are tests for partition replica listener. */
 @ExtendWith(ConfigurationExtension.class)
-public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstractTest {
+public class DefaultTablePartitionReplicaProcessorIndexLockingTest extends IgniteAbstractTest {
     private static final int PART_ID = 0;
     private static final int TABLE_ID = 1;
     private static final int PK_INDEX_ID = 1;
-    private static final int ZONE_ID = 2;
+    private static final int HASH_INDEX_ID = 2;
+    private static final int SORTED_INDEX_ID = 3;
     private static final IndexLocker.PartitionIndexId PK_INDEX_CONTEXT_ID
             = new IndexLocker.PartitionIndexId(PART_ID, PK_INDEX_ID);
+    private static final IndexLocker.PartitionIndexId HASH_INDEX_CONTEXT_ID
+            = new IndexLocker.PartitionIndexId(PART_ID, HASH_INDEX_ID);
+    private static final IndexLocker.PartitionIndexId SORTED_INDEX_CONTEXT_ID
+            = new IndexLocker.PartitionIndexId(PART_ID, SORTED_INDEX_ID);
+    private static final int ZONE_ID = 4;
     private static final UUID TRANSACTION_ID = TestTransactionIds.newTransactionId();
     private static final HybridClock CLOCK = new HybridClockImpl();
     private static final ClockService CLOCK_SERVICE = new TestClockService(CLOCK);
@@ -150,11 +161,14 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
     private static final PartitionReplicationMessagesFactory TABLE_MESSAGES_FACTORY = new PartitionReplicationMessagesFactory();
     private static final TestMvPartitionStorage TEST_MV_PARTITION_STORAGE = new TestMvPartitionStorage(PART_ID);
 
+    static final UUID LOCAL_NODE_ID = new UUID(0, 0);
+
     private static SchemaDescriptor schemaDescriptor;
     private static KvMarshaller<Integer, Integer> kvMarshaller;
     private static Lazy<TableSchemaAwareIndexStorage> pkStorage;
-    private static PartitionReplicaListener partitionReplicaListener;
+    private static DefaultTablePartitionReplicaProcessor tablePartitionReplicaProcessor;
     private static ColumnsExtractor row2HashKeyConverter;
+    private static ColumnsExtractor row2SortKeyConverter;
 
     @InjectConfiguration
     private static ReplicationConfiguration replicationConfiguration;
@@ -176,27 +190,45 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
 
         row2HashKeyConverter = BinaryRowConverter.keyExtractor(schemaDescriptor);
 
-        StorageSortedIndexDescriptor pkIndexDescriptor = new StorageSortedIndexDescriptor(
+        StorageHashIndexDescriptor pkIndexDescriptor = new StorageHashIndexDescriptor(
                 PK_INDEX_ID,
-                List.of(new StorageSortedIndexColumnDescriptor("ID", NativeTypes.INT32, false, true, false)),
+                List.of(new StorageHashIndexColumnDescriptor("ID", NativeTypes.INT32, false)),
                 false
         );
 
-        TableSchemaAwareIndexStorage pkSortedIndexStorage = new TableSchemaAwareIndexStorage(
+        TableSchemaAwareIndexStorage hashIndexStorage = new TableSchemaAwareIndexStorage(
                 PK_INDEX_ID,
-                new TestSortedIndexStorage(PART_ID, pkIndexDescriptor),
+                new TestHashIndexStorage(PART_ID, pkIndexDescriptor),
                 row2HashKeyConverter
         );
+        pkStorage = new Lazy<>(() -> hashIndexStorage);
 
-        pkStorage = new Lazy<>(() -> pkSortedIndexStorage);
+        IndexLocker pkLocker = new HashIndexLocker(PK_INDEX_ID, PART_ID, true, LOCK_MANAGER, row2HashKeyConverter);
+        IndexLocker hashIndexLocker = new HashIndexLocker(HASH_INDEX_ID, PART_ID, false, LOCK_MANAGER, row2HashKeyConverter);
 
-        IndexLocker pkLocker = new SortedIndexLocker(
-                PK_INDEX_ID,
+        BinaryTupleSchema rowSchema = BinaryTupleSchema.createRowSchema(schemaDescriptor);
+        BinaryTupleSchema keySchema = BinaryTupleSchema.createKeySchema(schemaDescriptor);
+
+        row2SortKeyConverter = new BinaryRowConverter(rowSchema, keySchema);
+
+        TableSchemaAwareIndexStorage sortedIndexStorage = new TableSchemaAwareIndexStorage(
+                SORTED_INDEX_ID,
+                new TestSortedIndexStorage(PART_ID,
+                        new StorageSortedIndexDescriptor(
+                                SORTED_INDEX_ID,
+                                List.of(new StorageSortedIndexColumnDescriptor("val", NativeTypes.INT32, false, true, false)),
+                                false
+                        )),
+                row2SortKeyConverter
+        );
+
+        IndexLocker sortedIndexLocker = new SortedIndexLocker(
+                SORTED_INDEX_ID,
                 PART_ID,
                 LOCK_MANAGER,
-                (SortedIndexStorage) pkSortedIndexStorage.storage(),
-                row2HashKeyConverter,
-                true
+                (SortedIndexStorage) sortedIndexStorage.storage(),
+                row2SortKeyConverter,
+                false
         );
 
         DummySchemaManagerImpl schemaManager = new DummySchemaManagerImpl(schemaDescriptor);
@@ -208,12 +240,10 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
 
         TestPartitionDataStorage partitionDataStorage = new TestPartitionDataStorage(TABLE_ID, PART_ID, TEST_MV_PARTITION_STORAGE);
 
-        CatalogService catalogService = mock(CatalogService.class);
         Catalog catalog = mock(Catalog.class);
-
-        when(catalogService.catalog(anyInt())).thenReturn(catalog);
+        CatalogService catalogService = mock(CatalogService.class);
         when(catalogService.activeCatalog(anyLong())).thenReturn(catalog);
-        when(catalogService.catalogReadyFuture(anyInt())).thenReturn(nullCompletedFuture());
+        when(catalogService.catalog(anyInt())).thenReturn(catalog);
 
         CatalogTableDescriptor tableDescriptor = mock(CatalogTableDescriptor.class);
         when(tableDescriptor.latestSchemaVersion()).thenReturn(schemaDescriptor.version());
@@ -228,7 +258,7 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
 
         InternalClusterNode localNode = DummyInternalTableImpl.LOCAL_NODE;
 
-        partitionReplicaListener = new PartitionReplicaListener(
+        tablePartitionReplicaProcessor = new DefaultTablePartitionReplicaProcessor(
                 TEST_MV_PARTITION_STORAGE,
                 mockRaftClient,
                 newTxManager(),
@@ -236,9 +266,16 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
                 Runnable::run,
                 new ZonePartitionId(ZONE_ID, PART_ID),
                 TABLE_ID,
-                () -> Int2ObjectMaps.singleton(pkLocker.id(), pkLocker),
+                () -> Int2ObjectMap.ofEntries(
+                        Int2ObjectMap.entry(pkLocker.id(), pkLocker),
+                        Int2ObjectMap.entry(hashIndexLocker.id(), hashIndexLocker),
+                        Int2ObjectMap.entry(sortedIndexLocker.id(), sortedIndexLocker)
+                ),
                 pkStorage,
-                Int2ObjectMaps::emptyMap,
+                () -> Int2ObjectMap.ofEntries(
+                        Int2ObjectMap.entry(sortedIndexLocker.id(), sortedIndexStorage),
+                        Int2ObjectMap.entry(hashIndexLocker.id(), hashIndexStorage)
+                ),
                 CLOCK_SERVICE,
                 safeTime,
                 mock(TransactionStateResolver.class),
@@ -294,22 +331,18 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
         return txManager;
     }
 
-    private static ReplicaPrimacy validRwPrimacy() {
-        return ReplicaPrimacy.forPrimaryReplicaRequest(1);
-    }
-
-    @BeforeEach
-    public void beforeTest() {
-        ((TestSortedIndexStorage) pkStorage.get().storage()).clear();
-        TEST_MV_PARTITION_STORAGE.clear();
-
-        LOCK_MANAGER.releaseAll(TRANSACTION_ID);
-    }
-
     private static LockManager lockManager() {
         HeapLockManager lockManager = HeapLockManager.smallInstance();
         lockManager.start(new WaitDieDeadlockPreventionPolicy());
         return lockManager;
+    }
+
+    @BeforeEach
+    public void beforeTest() {
+        ((TestHashIndexStorage) pkStorage.get().storage()).clear();
+        TEST_MV_PARTITION_STORAGE.clear();
+
+        LOCK_MANAGER.releaseAll(TRANSACTION_ID);
     }
 
     /** Verifies the mode in which the lock was acquired on the index key for a particular operation. */
@@ -370,7 +403,7 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
                 throw new AssertionError("Unexpected operation type: " + arg.type);
         }
 
-        CompletableFuture<?> fut = partitionReplicaListener.process(request, validRwPrimacy(), LOCAL_NODE_ID);
+        CompletableFuture<?> fut = tablePartitionReplicaProcessor.process(request, replicaPrimacy(), LOCAL_NODE_ID);
 
         await(fut);
 
@@ -378,13 +411,27 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
                 locks(),
                 allOf(
                         hasItem(lockThat(
-                                arg.expectedLockOnSortedPk + " on sorted pk index",
+                                arg.expectedLockOnUniqueHash + " on unique hash index",
                                 lock -> Objects.equals(PK_INDEX_CONTEXT_ID, lock.lockKey().contextId())
                                         && row2HashKeyConverter.extractColumns(testBinaryRow).byteBuffer().equals(lock.lockKey().key())
-                                        && lock.lockMode() == arg.expectedLockOnSortedPk
+                                        && lock.lockMode() == arg.expectedLockOnUniqueHash
+                        )),
+                        hasItem(lockThat(
+                                arg.expectedLockOnNonUniqueHash + " on non unique hash index",
+                                lock -> Objects.equals(HASH_INDEX_CONTEXT_ID, lock.lockKey().contextId())
+                                        && lock.lockMode() == arg.expectedLockOnNonUniqueHash
+                        )),
+                        hasItem(lockThat(
+                                arg.expectedLockOnSort + " on sorted index",
+                                lock -> Objects.equals(SORTED_INDEX_CONTEXT_ID, lock.lockKey().contextId())
+                                        && lock.lockMode() == arg.expectedLockOnSort
                         ))
                 )
         );
+    }
+
+    private ReplicaPrimacy replicaPrimacy() {
+        return ReplicaPrimacy.forPrimaryReplicaRequest(HybridTimestamp.MIN_VALUE.longValue());
     }
 
     /** Verifies the mode in which the lock was acquired on the index key for a particular operation. */
@@ -449,7 +496,7 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
                 throw new AssertionError("Unexpected operation type: " + arg.type);
         }
 
-        CompletableFuture<?> fut = partitionReplicaListener.process(request, validRwPrimacy(), LOCAL_NODE_ID);
+        CompletableFuture<?> fut = tablePartitionReplicaProcessor.process(request, replicaPrimacy(), LOCAL_NODE_ID);
 
         await(fut);
 
@@ -458,10 +505,22 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
                     locks(),
                     allOf(
                             hasItem(lockThat(
-                                    arg.expectedLockOnSortedPk + " on sorted pk index",
+                                    arg.expectedLockOnUniqueHash + " on unique hash index",
                                     lock -> Objects.equals(PK_INDEX_CONTEXT_ID, lock.lockKey().contextId())
                                             && row2HashKeyConverter.extractColumns(row).byteBuffer().equals(lock.lockKey().key())
-                                            && lock.lockMode() == arg.expectedLockOnSortedPk
+                                            && lock.lockMode() == arg.expectedLockOnUniqueHash
+                            )),
+                            hasItem(lockThat(
+                                    arg.expectedLockOnNonUniqueHash + " on non unique hash index",
+                                    lock -> Objects.equals(HASH_INDEX_CONTEXT_ID, lock.lockKey().contextId())
+                                            && row2HashKeyConverter.extractColumns(row).byteBuffer().equals(lock.lockKey().key())
+                                            && lock.lockMode() == arg.expectedLockOnNonUniqueHash
+                            )),
+                            hasItem(lockThat(
+                                    arg.expectedLockOnSort + " on sorted index",
+                                    lock -> Objects.equals(SORTED_INDEX_CONTEXT_ID, lock.lockKey().contextId())
+                                            && row2SortKeyConverter.extractColumns(row).byteBuffer().equals(lock.lockKey().key())
+                                            && lock.lockMode() == arg.expectedLockOnSort
                             ))
                     )
             );
@@ -470,24 +529,24 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
 
     private static Iterable<ReadWriteTestArg> readWriteSingleTestArguments() {
         return List.of(
-                new ReadWriteTestArg(RW_DELETE, LockMode.SIX),
-                new ReadWriteTestArg(RW_DELETE_EXACT, LockMode.SIX),
-                new ReadWriteTestArg(RW_INSERT, LockMode.X),
-                new ReadWriteTestArg(RW_UPSERT, LockMode.X),
-                new ReadWriteTestArg(RW_REPLACE_IF_EXIST, LockMode.X),
+                new ReadWriteTestArg(RW_DELETE, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_DELETE_EXACT, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_INSERT, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_UPSERT, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_REPLACE_IF_EXIST, LockMode.X, LockMode.IX, LockMode.IX),
 
-                new ReadWriteTestArg(RW_GET_AND_DELETE, LockMode.SIX),
-                new ReadWriteTestArg(RW_GET_AND_REPLACE, LockMode.X),
-                new ReadWriteTestArg(RW_GET_AND_UPSERT, LockMode.X)
+                new ReadWriteTestArg(RW_GET_AND_DELETE, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_GET_AND_REPLACE, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_GET_AND_UPSERT, LockMode.X, LockMode.IX, LockMode.IX)
         );
     }
 
     private static Iterable<ReadWriteTestArg> readWriteMultiTestArguments() {
         return List.of(
-                new ReadWriteTestArg(RW_DELETE_ALL, LockMode.SIX),
-                new ReadWriteTestArg(RW_DELETE_EXACT_ALL, LockMode.SIX),
-                new ReadWriteTestArg(RW_INSERT_ALL, LockMode.X),
-                new ReadWriteTestArg(RW_UPSERT_ALL, LockMode.X)
+                new ReadWriteTestArg(RW_DELETE_ALL, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_DELETE_EXACT_ALL, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_INSERT_ALL, LockMode.X, LockMode.IX, LockMode.IX),
+                new ReadWriteTestArg(RW_UPSERT_ALL, LockMode.X, LockMode.IX, LockMode.IX)
         );
     }
 
@@ -527,14 +586,20 @@ public class PartitionReplicaListenerSortedIndexLockingTest extends IgniteAbstra
 
     static class ReadWriteTestArg {
         private final RequestType type;
-        private final LockMode expectedLockOnSortedPk;
+        private final LockMode expectedLockOnUniqueHash;
+        private final LockMode expectedLockOnNonUniqueHash;
+        private final LockMode expectedLockOnSort;
 
         ReadWriteTestArg(
                 RequestType type,
-                LockMode expectedLockOnSortedPk
+                LockMode expectedLockOnUniqueHash,
+                LockMode expectedLockOnNonUniqueHash,
+                LockMode expectedLockOnSort
         ) {
             this.type = type;
-            this.expectedLockOnSortedPk = expectedLockOnSortedPk;
+            this.expectedLockOnUniqueHash = expectedLockOnUniqueHash;
+            this.expectedLockOnNonUniqueHash = expectedLockOnNonUniqueHash;
+            this.expectedLockOnSort = expectedLockOnSort;
         }
 
         @Override
